refactor: avoid adding `typing.CompliantNamespace`
https://github.com/narwhals-dev/narwhals/pull/2149#discussion_r1981760892
